% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{ggiNEXTBetaDiv}
\alias{ggiNEXTBetaDiv}
\title{ggplot for Beta diversity}
\usage{
ggiNEXTBetaDiv(output, type = "B", scale = "free", main = NULL, transp = 0.4)
}
\arguments{
\item{output}{the output from iNEXTBetaDiv}

\item{type}{selection of plot type : \code{type = 'B'} for plotting the gamma, alpha, and beta diversity ; \code{type = 'D'} for plotting 4 turnover dissimilarities.}

\item{scale}{Are scales shared across all facets (\code{"fixed"}), or do they vary across rows (\code{"free_x"}), columns (\code{"free_y"}), or both rows and columns (\code{"free"})? Default is \code{"free"}.}

\item{main}{The title of the plot.}

\item{transp}{a value between 0 and 1 controlling transparency. \code{transp = 0} is completely transparent, default is 0.4.}
}
\value{
a figure for Beta diversity or dissimilarity diversity.
}
\description{
\code{ggiNEXTBetaDiv}: ggplot for Interpolation and extrapolation of Beta diversity with order q
}
\examples{
## abundance data
data(beetle_abu)
output1 = iNEXTBetaDiv(data = beetle_abu, datatype = 'abundance', level = seq(0.8, 1, 0.05), 
                       nboot = 20)

ggiNEXTBetaDiv(output1, type = 'B', scale = 'free', main = NULL, transp = 0.4)
ggiNEXTBetaDiv(output1, type = 'D', scale = 'free', main = NULL, transp = 0.4)


## incidence data
data(beetle_inc)
output2 = iNEXTBetaDiv(data = beetle_inc, datatype = 'incidence_raw', level = seq(0.8, 1, 0.05),
                       nboot = 20, conf = 0.95)

ggiNEXTBetaDiv(output2, type = 'B', scale = 'free', main = NULL, transp = 0.4)
ggiNEXTBetaDiv(output2, type = 'D', scale = 'free', main = NULL, transp = 0.4)


#' ## abundance data & size-based
data(beetle_abu)
output3 = iNEXTBetaDiv(data = beetle_abu, datatype = 'abundance', base = 'size',
                       nboot = 20, conf = 0.95)
ggiNEXTBetaDiv(output3, scale = 'free', main = NULL, transp = 0.4)


## incidence data & size-based
data(beetle_inc)
output4 = iNEXTBetaDiv(data = beetle_inc, datatype = 'incidence_raw', base = 'size',
                       nboot = 20, conf = 0.95)
ggiNEXTBetaDiv(output4, type = 'B', scale = 'free', main = NULL, transp = 0.4)


}
