% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{iNEXTBetaDiv}
\alias{iNEXTBetaDiv}
\title{iNterpolation and EXTrapolation for Beta diversity}
\usage{
iNEXTBetaDiv(
  data,
  q = c(0, 1, 2),
  datatype = "abundance",
  base = "coverage",
  level = NULL,
  nboot = 20,
  conf = 0.95
)
}
\arguments{
\item{data}{(a) For \code{datatype = "abundance"}, data can be input as a \code{matrix/data.frame} (species by assemblages), or a \code{list} of \code{matrices/data.frames}, each matrix represents species-by-assemblages abundance matrix; see Note 1 for examples.\cr
(b) For \code{datatype = "incidence_raw"}, data can be input as a \code{list} of \code{matrices/data.frames}, each matrix represents species-by-sampling units; see Note 2 for an example.}

\item{q}{a numerical vector specifying the diversity orders. Default is c(0, 1, 2).}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}) with all entries being \code{0} (non-detection) or \code{1} (detection).}

\item{base}{Sample-sized-based rarefaction and extrapolation for gamma and alpha diversity (\code{base = "size"}) or coverage-based rarefaction and extrapolation for gamma, alpha and beta diversity (\code{base = "coverage"}). Default is \code{base = "coverage"}.}

\item{level}{A numerical vector specifying the particular value of sample coverage (between 0 and 1 when \code{base = "coverage"}) or sample size (\code{base = "size"}). \code{level = 1} (\code{base = "coverage"}) means complete coverage (the corresponding diversity represents asymptotic diversity).\cr
If \code{base = "size"} and \code{level = NULL}, then this function computes the gamma and alpha diversity estimates up to double the reference sample size. If \code{base = "coverage"} and \code{level = NULL}, then this function computes the gamma and alpha diversity estimates up to one (for \code{q = 1, 2}) or up to the coverage of double the reference sample size (for \code{q = 0});\cr 
the corresponding beta diversity is computed up to the same maximum coverage as the alpha diversity.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter \code{0} to skip the bootstrap procedures. Default is \code{20}. If more accurate results are required, set \code{nboot = 100} (or \code{200}).}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is \code{0.95}.}
}
\value{
A list of seven lists with three-diversity and four-dissimilarity.
}
\description{
\code{iNEXTBetaDiv}: Interpolation and extrapolation of Beta diversity with order q
}
\examples{
## abundance data & coverage-based
data(beetle_abu)
output1 = iNEXTBetaDiv(data = beetle_abu, datatype = 'abundance',
                       nboot = 20, conf = 0.95)
output1


## incidence data & coverage-based
data(beetle_inc)
output2 = iNEXTBetaDiv(data = beetle_inc, datatype = 'incidence_raw',
                       nboot = 20, conf = 0.95)
output2


## abundance data & size-based
data(beetle_abu)
output3 = iNEXTBetaDiv(data = beetle_abu, datatype = 'abundance', base = 'size',
                       nboot = 20, conf = 0.95)
output3


## incidence data & size-based
data(beetle_inc)
output4 = iNEXTBetaDiv(data = beetle_inc, datatype = 'incidence_raw', base = 'size',
                       nboot = 20, conf = 0.95)
output4


}
