% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{iNEXTBetaDiv}
\alias{iNEXTBetaDiv}
\title{iNterpolation and EXTrapolation for Beta diversity}
\usage{
iNEXTBetaDiv(
  data,
  q = c(0, 1, 2),
  datatype = "abundance",
  level,
  nboot = 20,
  conf = 0.95
)
}
\arguments{
\item{data}{(a) For \code{datatype = "abundance"}, data can be input as a matrix/data.frame (a region with species by assemblages), or a list (multiple regions) of several matrices/data.frames (species by assemblages).\cr
(b) For \code{datatype = "incidence_raw"}, data can be input as a list of several lists (regions)  contained matrices/data.frames (species by sampling units).}

\item{q}{a numerical vector specifying the diversity orders. Default is c(0, 1, 2).}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}) with all entries being 0 (non-detection) or 1 (detection).}

\item{level}{A numerical vector specifying the particular sample coverages (between 0 and 1). Calculating asymptotic diversity if \code{level = 1}. \cr}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Enter 0 to skip the bootstrap procedures. Default is 20.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is 0.95.}
}
\value{
A list of seven lists with three-diversity and four-dissimilarity.
}
\description{
\code{iNEXTBetaDiv}: Interpolation and extrapolation of Beta diversity with order q
}
\examples{
## abundance data
data(beetle_abu)
output1 = iNEXTBetaDiv(data = beetle_abu, datatype = 'abundance', level = seq(0.8, 1, 0.05), 
                       nboot = 20, conf = 0.95)
output1


## incidence data
data(beetle_inc)
output2 = iNEXTBetaDiv(data = beetle_inc, datatype = 'incidence_raw', level = seq(0.8, 1, 0.05),
                       nboot = 20, conf = 0.95)
output2



}
